requirements.txt
matplotlib~=3.9.2
numpy~=2.1.2
pandas~=2.2.3
requests~=2.32.3
streamlit~=1.39.0
psychopg2~=2.9.10
sqlalchemy~=2.0.36
python-dotenv~=1.0.1